address native notes shell review feedback
