+++ packages/api/src/__tests__/initApi.test.ts
@@ -1,23 +1,18 @@
-import { beforeAll, describe, expect, test } from 'vitest';
+import { describe, expect, test } from 'vitest';
 
 import { toInitData } from '../client/initApi';
-import { internalConfigureClient } from '../client/urbit';
 import type * as ub from '../urbit';
 import rawGroupsInit6 from './fixtures/groupsInit5.json';
 
 const groupsInit6 = rawGroupsInit6 as unknown as ub.GroupsInit6;
+const currentUserId = '~solfer-magfed';
+const toTestInitData = (response: ub.GroupsInit7) =>
+  toInitData(response, { currentUserId });
 
 describe('toInitData', () => {
-  beforeAll(() => {
-    internalConfigureClient({
-      shipName: 'solfer-magfed',
-      shipUrl: 'http://localhost',
-    });
-  });
-
   describe('v6 format support', () => {
     test('processes v6 init with foreigns', () => {
-      const result = toInitData(groupsInit6);
+      const result = toTestInitData(groupsInit6);
 
       expect(result).toBeDefined();
       expect(result.groups).toBeDefined();
@@ -27,7 +22,7 @@ describe('toInitData', () => {
     });
 
     test('converts v7 groups correctly', () => {
-      const result = toInitData(groupsInit6);
+      const result = toTestInitData(groupsInit6);
 
       const groups = result.groups;
       expect(groups.length).toBeGreaterThan(0);
@@ -45,7 +40,7 @@ describe('toInitData', () => {
     });
 
     test('converts foreigns to unjoined groups', () => {
-      const result = toInitData(groupsInit6);
+      const result = toTestInitData(groupsInit6);
 
       expect(result.unjoinedGroups).toBeDefined();
       expect(result.unjoinedGroups.length).toBeGreaterThan(0);
@@ -58,7 +53,7 @@ describe('toInitData', () => {
     });
 
     test('extracts channel readers from v7 groups', () => {
-      const result = toInitData(groupsInit6);
+      const result = toTestInitData(groupsInit6);
 
       expect(result.channelPerms).toBeDefined();
       expect(result.channelPerms.length).toBeGreaterThan(0);
@@ -76,14 +71,14 @@ describe('toInitData', () => {
         'notes/~test-ship/native-notes',
       ];
 
-      const result = toInitData(response);
+      const result = toTestInitData(response);
 
       expect(result.joinedGroupChannels).toContain(
         'notes/~test-ship/native-notes'
       );
     });
 
-    test('treats auto-joined notes listings as joined for old backends', () => {
+    test('does not treat notes listings as joined without active channel state', () => {
       const response = structuredClone(groupsInit6);
       const group = response.groups['~test-ship/test-group'];
       group['active-channels'] = [];
@@ -100,63 +95,15 @@ describe('toInitData', () => {
         join: true,
       };
 
-      const result = toInitData(response);
+      const result = toTestInitData(response);
 
-      expect(result.joinedGroupChannels).toContain(
+      expect(result.joinedGroupChannels).not.toContain(
         'notes/~test-ship/native-notes'
       );
     });
 
-    test('does not treat non-notes listings as joined without active channel state', () => {
-      const response = structuredClone(groupsInit6);
-      const group = response.groups['~test-ship/test-group'];
-      group['active-channels'] = [];
-      group.channels['heap/~test-ship/announcements'] = {
-        added: 1692473215572,
-        meta: {
-          image: '',
-          title: 'Announcements',
-          cover: '',
-          description: '',
-        },
-        section: 'default',
-        readers: [],
-        join: true,
-      };
-
-      const result = toInitData(response);
-
-      expect(result.joinedGroupChannels).not.toContain(
-        'heap/~test-ship/announcements'
-      );
-    });
-
-    test('does not treat unreadable notes listings as joined', () => {
-      const response = structuredClone(groupsInit6);
-      const group = response.groups['~test-ship/test-group'];
-      group['active-channels'] = [];
-      group.channels['notes/~test-ship/private-notes'] = {
-        added: 1692473215572,
-        meta: {
-          image: '',
-          title: 'Private notes',
-          cover: '',
-          description: '',
-        },
-        section: 'default',
-        readers: ['not-my-role'],
-        join: true,
-      };
-
-      const result = toInitData(response);
-
-      expect(result.joinedGroupChannels).not.toContain(
-        'notes/~test-ship/private-notes'
-      );
-    });
-
     test('filters valid invites from foreigns', () => {
-      const result = toInitData(groupsInit6);
+      const result = toTestInitData(groupsInit6);
 
       const foreignGroup = result.unjoinedGroups[0];
       // Should have invite because valid=true in fixture
